git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@8764 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/quantize.c
@@ -1401,7 +1401,7 @@ static RealPixelPacket **AcquirePixelThreadSet(const size_t count)
   size_t
     number_threads;
 
-  number_threads=GetMagickResourceLimit(ThreadResource);
+  number_threads=(size_t) GetMagickResourceLimit(ThreadResource);
   pixels=(RealPixelPacket **) AcquireQuantumMemory(number_threads,
     sizeof(*pixels));
   if (pixels == (RealPixelPacket **) NULL)
