+++ magick/statistic.c
@@ -160,7 +160,7 @@ static MagickPixelPacket **AcquirePixelThreadSet(const Image *image,
     length,
     number_threads;
 
-  number_threads=GetMagickResourceLimit(ThreadResource);
+  number_threads=(size_t) GetMagickResourceLimit(ThreadResource);
   pixels=(MagickPixelPacket **) AcquireQuantumMemory(number_threads,
     sizeof(*pixels));
   if (pixels == (MagickPixelPacket **) NULL)
@@ -2016,7 +2016,7 @@ static PixelList **AcquirePixelListThreadSet(const size_t width,
   size_t
     number_threads;
 
-  number_threads=GetMagickResourceLimit(ThreadResource);
+  number_threads=(size_t) GetMagickResourceLimit(ThreadResource);
   pixel_list=(PixelList **) AcquireQuantumMemory(number_threads,
     sizeof(*pixel_list));
   if (pixel_list == (PixelList **) NULL)
